Move reraise from utils to backend
