+++ lmfdb/backend/utils.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function, absolute_import
-import re
+import sys, re
 
 from psycopg2.sql import SQL, Identifier, Placeholder
 
@@ -199,3 +199,20 @@ class DelayCommit(object):
             self.obj.conn.commit()
         if exc_type is not None:
             self.obj.conn.rollback()
+
+# Reraise an exception, possibly with a different message, type, or traceback.
+if sys.version_info.major < 3:  # Python 2?
+    # Using exec avoids a SyntaxError in Python 3.
+    exec("""def reraise(exc_type, exc_value, exc_traceback=None):
+            # Reraise an exception, possibly with a different message, type, or traceback.
+                raise exc_type, exc_value, exc_traceback""")
+else:
+    def reraise(exc_type, exc_value, exc_traceback=None):
+        """
+        Reraise an exception, possibly with a different message, type, or traceback.
+        """
+        if exc_value is None:
+            exc_value = exc_type()
+        if exc_value.__traceback__ is not exc_traceback:
+            raise exc_value.with_traceback(exc_traceback)
+        raise exc_value
