+++ lmfdb/utils/__init__.py
@@ -42,7 +42,6 @@ __all__ = ['request', 'make_response', 'flash', 'url_for', 'render_template',
            'Configuration',
            'names_and_urls', 'name_and_object_from_url',
            'datetime_to_timestamp_in_ms', 'timestamp_in_ms_to_datetime',
-           'reraise',
            'TraceHash', 'TraceHashClass']
 
 from flask import (request, make_response, flash, url_for,
@@ -89,5 +88,4 @@ from .downloader import Downloader
 from .display_stats import formatters, proportioners, totaler, StatsDisplay
 from .config import Configuration
 from .names_and_urls import names_and_urls, name_and_object_from_url
-from .reraise import reraise
 from .trace_hash import TraceHash, TraceHashClass
