--- lmfdb/utils/reraise.py
@@ -1,18 +1,0 @@
-import sys
-
-# Reraise an exception, possibly with a different message, type, or traceback.
-if sys.version_info.major < 3:  # Python 2?
-    # Using exec avoids a SyntaxError in Python 3.
-    exec("""def reraise(exc_type, exc_value, exc_traceback=None):
-            # Reraise an exception, possibly with a different message, type, or traceback.
-                raise exc_type, exc_value, exc_traceback""")
-else:
-    def reraise(exc_type, exc_value, exc_traceback=None):
-        """
-        Reraise an exception, possibly with a different message, type, or traceback.
-        """
-        if exc_value is None:
-            exc_value = exc_type()
-        if exc_value.__traceback__ is not exc_traceback:
-            raise exc_value.with_traceback(exc_traceback)
-        raise exc_value
